move eerors in own file for better loading
